Move superblock code to neuralnet directory
