+++ src/neuralnet/superblock.cpp
